session.c: fix compiler warning

+++ src/session.c
@@ -95,7 +95,7 @@ LIBSSH2_REALLOC_FUNC(libssh2_default_realloc)
 static int
 banner_receive(LIBSSH2_SESSION * session)
 {
-    int ret;
+    ssize_t ret;
     size_t banner_len;
 
     if(session->banner_TxRx_state == libssh2_NB_state_idle) {
@@ -121,11 +121,11 @@ banner_receive(LIBSSH2_SESSION * session)
             if(session->api_block_mode || (ret != -EAGAIN))
                 /* ignore EAGAIN when non-blocking */
                 _libssh2_debug((session, LIBSSH2_TRACE_SOCKET,
-                               "Error recving %d bytes: %d", 1, -ret));
+                               "Error recving %d bytes: %d", 1, (int)-ret));
         }
         else
             _libssh2_debug((session, LIBSSH2_TRACE_SOCKET,
-                           "Recved %d bytes banner", ret));
+                           "Recved %d bytes banner", (int)ret));
 
         if(ret < 0) {
             if(ret == -EAGAIN) {
